fix: 신입미션 Embed 미생성 버그 수정 및 moco-db 타입 캐스팅 보완
- removeInvalidMissions에서 Discord API 일시 오류 시 미션이 LEFT+hidden으로
  영구 손상되는 치명적 버그 수정 (DiscordAPIError 10007만 탈퇴로 판정)
- registerMissingMembers 예외 시 embed 갱신 전체 중단되는 문제 try-catch로 해결
- Embed description 4096자 초과 시 Discord API 에러 방지 (truncate 처리)
- setColor에 유효하지 않은 색상 문자열 전달 시 throw 방지 (안전 파싱 fallback)
- moco-db upsert 쿼리에 ::int, ::numeric 타입 캐스팅 추가

+++ apps/api/src/newbie/infrastructure/moco-db.repository.ts
@@ -64,16 +64,16 @@ export class MocoDbRepository {
       `
       INSERT INTO ${schemaPrefix}"${tableName}"
         ("guildId", "hunterId", "date", "channelMinutes", "sessionCount", "uniqueNewbieCount", "score")
-      VALUES ($1, $2, $3, $4, $5, $6,
-        ($5 * $7) + ($4 * $8) + ($6 * $9)
+      VALUES ($1, $2, $3, $4::int, $5::int, $6::int,
+        ($5::int * $7::numeric) + ($4::int * $8::numeric) + ($6::int * $9::numeric)
       )
       ON CONFLICT ("guildId", "hunterId", "date") DO UPDATE SET
         "channelMinutes"    = "${tableName}"."channelMinutes"    + EXCLUDED."channelMinutes",
         "sessionCount"      = "${tableName}"."sessionCount"      + EXCLUDED."sessionCount",
         "uniqueNewbieCount" = "${tableName}"."uniqueNewbieCount" + EXCLUDED."uniqueNewbieCount",
-        "score" = (("${tableName}"."sessionCount" + EXCLUDED."sessionCount") * $7)
-                + (("${tableName}"."channelMinutes" + EXCLUDED."channelMinutes") * $8)
-                + (("${tableName}"."uniqueNewbieCount" + EXCLUDED."uniqueNewbieCount") * $9)
+        "score" = (("${tableName}"."sessionCount" + EXCLUDED."sessionCount") * $7::numeric)
+                + (("${tableName}"."channelMinutes" + EXCLUDED."channelMinutes") * $8::numeric)
+                + (("${tableName}"."uniqueNewbieCount" + EXCLUDED."uniqueNewbieCount") * $9::numeric)
       `,
       [
         guildId,
